pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "selfops"                      # Must be unique on PyPI
version = "0.1.2"
description = "A simple CLI built using Typer"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Tirth Sharma and Team of SeflOps.", email = "sharmatirth09012@gmail.com" }]
dependencies = [
    "annotated-doc==0.0.3",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "bidict==0.23.1",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "docker==7.1.0",
    "fastapi==0.120.4",
    "h11==0.16.0",
    "idna==3.11",
    "mdurl==0.1.2",
    "pycparser==2.23",
    "pydantic_core==2.41.4",
    "Pygments==2.19.2",
    "PyJWT==2.10.1",
    "python-decouple==3.8",
    "python-engineio==4.12.3",
    "python-socketio==5.14.3",
    "requests==2.32.5",
    "shellingham==1.5.4",
    "simple-websocket==1.1.0",
    "sniffio==1.3.1",
    "starlette==0.49.1",
    "typer==0.20.0",
    "typing-inspection==0.4.2",
    "typing_extensions==4.15.0",
    "urllib3==2.5.0",
    "wsproto==1.2.0",

    "rich>=13.7.0,<15.0.0",
    "pydantic>=2.0,<2.13",
    "cryptography>=42.0,<47.0",
    "certifi>=2024.7,<2026.0",
    "markdown-it-py>=3.0,<4.1",
]


[tool.setuptools]
packages = ["core", "commands", "utils"]
py-modules = ["main"]

[project.scripts]
selfops = "main:app"